requirements.txt
flask
flask-cors
yagmail
requests
python-dotenv
gunicorn
